app works, dag does not
