+++ mpx_dashboard/population.csv
@@ -1,0 +1,119 @@
+country,population
+Andorra,77265.0
+United Arab Emirates,9890400.0
+Argentina,44938712.0
+Austria,8858775.0
+Australia,25499881.0
+Aruba,106766.0
+Bosnia And Herzegovina,3280815.0
+Barbados,287371.0
+Belgium,11455519.0
+Bulgaria,7000039.0
+Bahrain,1701583.0
+Benin,12123198.0
+Bermuda,62273.0
+Bolivia,11673029.0
+Brazil,212559409.0
+Bahamas,393248.0
+Canada,37742157.0
+Democratic Republic Of The Congo,89561404.0
+Central African Republic,4829764.0
+Republic of Congo,5518092.0
+Switzerland,8544527.0
+Chile,17574003.0
+Cameroon,26545864.0
+China,1439323774.0
+Colombia,50882884.0
+Costa Rica,5094114.0
+Cuba,11326616.0
+Curaçao,164100.0
+Cyprus,875899.0
+Czech Republic,10649800.0
+Germany,83019213.0
+Denmark,5806081.0
+Dominican Republic,10847904.0
+Ecuador,17643060.0
+Estonia,1324820.0
+Egypt,102334403.0
+Spain,46937060.0
+Finland,5517919.0
+France,67012883.0
+Guadeloupe,416474.0
+England,53012456.0
+Northern Ireland,1885189.0
+Scotland,5454238.0
+Wales,3136383.0
+Georgia,3989175.0
+Ghana,31072945.0
+Gibraltar,33691.0
+Greenland,56772.0
+Greece,10724599.0
+Guatemala,17915567.0
+Guam,168783.0
+Guyana,786559.0
+Hong Kong,7496988.0
+Honduras,9904608.0
+Croatia,4076246.0
+Hungary,9772756.0
+Indonesia,273523621.0
+Ireland,4904240.0
+Israel,8655541.0
+India,1380004385.0
+Iran,83992953.0
+Iceland,356991.0
+Italy,60359546.0
+Jamaica,2961161.0
+Jordan,10203140.0
+Japan,126476458.0
+South Korea,51269183.0
+Cayman Islands,65720.0
+Lebanon,6825442.0
+Liberia,5057677.0
+Lithuania,2794184.0
+Luxembourg,613894.0
+Latvia,1919968.0
+Morocco,36910558.0
+Monaco,39244.0
+Moldova,4033963.0
+Montenegro,622182.0
+Martinique,375265.0
+Malta,493559.0
+Mexico,110991953.0
+Saint Martin (French part),2540916.0
+New Caledonia,285491.0
+Nigeria,206139587.0
+Netherlands,17282163.0
+Norway,5328212.0
+New Zealand,4822233.0
+Panama,4314768.0
+Peru,29381884.0
+Philippines,100979303.0
+Pakistan,220892331.0
+Poland,37972812.0
+Puerto Rico,2860840.0
+Portugal,10276617.0
+Paraguay,7132530.0
+Qatar,2881060.0
+Romania,19414458.0
+Serbia,6963764.0
+Russia,145934460.0
+Saudi Arabia,34813867.0
+Sudan,43849269.0
+Sweden,10230185.0
+Singapore,5850343.0
+Slovenia,2080908.0
+Slovakia,5450421.0
+Sierra Leone,7976985.0
+Somalia,15893219.0
+South Sudan,11193729.0
+El Salvador,6486201.0
+Thailand,69799978.0
+Turkey,82003882.0
+Taiwan,23816775.0
+Ukraine,43733759.0
+Uganda,45741000.0
+United States,331002647.0
+Uruguay,3473727.0
+Venezuela,28435943.0
+South Africa,59308690.0
+Zambia,18383956.0
