+++ mpx_dashboard/utils.py
@@ -1,0 +1,1390 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Sep 28 22:16:24 2022
+
+@author: nico
+"""
+rev_dict = lambda x: {v: k for k,v in x.items()}
+
+country_to_iso3 = {
+ 'Aruba': 'ABW', 
+ 'Afghanistan': 'AFG', 
+ 'Angola': 'AGO', 
+ 'Anguilla': 'AIA', 
+ 'Åland Islands': 'ALA', 
+ 'Albania': 'ALB', 
+ 'Andorra': 'AND', 
+ 'United Arab Emirates': 'ARE', 
+ 'Argentina': 'ARG', 
+ 'Armenia': 'ARM', 
+ 'American Samoa': 'ASM', 
+ 'Antarctica': 'ATA', 
+ 'French Southern Territories': 'ATF', 
+ 'Antigua and Barbuda': 'ATG', 
+ 'Antigua And Barbuda': 'ATG', 
+ 'Australia': 'AUS', 
+ 'Austria': 'AUT', 
+ 'Azerbaijan': 'AZE', 
+ 'Burundi': 'BDI', 
+ 'Belgium': 'BEL', 
+ 'Benin': 'BEN', 
+ 'Caribbean Netherlands': 'BES', 
+ 'Bonaire, Sint Eustatius and Saba': 'BES', 
+ 'Burkina Faso': 'BFA', 
+ 'Bangladesh': 'BGD', 
+ 'Bulgaria': 'BGR', 
+ 'Bahrain': 'BHR', 
+ 'Bahamas': 'BHS', 
+ 'Bosnia and Herzegovina': 'BIH', 
+ 'Bosnia And Herzegovina': 'BIH', 
+ 'Saint Barthélemy': 'BLM', 
+ 'Belarus': 'BLR', 
+ 'Belize': 'BLZ', 
+ 'Bermuda': 'BMU', 
+ 'Bolivia': 'BOL', 
+ 'Bolivia (Plurinational State of)': 'BOL', 
+ 'Brazil': 'BRA', 
+ 'Barbados': 'BRB', 
+ 'Brunei': 'BRN', 
+ 'Brunei Darussalam': 'BRN', 
+ 'Bhutan': 'BTN', 
+ 'Bouvet Island': 'BVT', 
+ 'Botswana': 'BWA', 
+ 'Central African Republic': 'CAF', 
+ 'Canada': 'CAN', 
+ 'Cocos Islands': 'CCK', 
+ 'Cocos (Keeling) Islands': 'CCK', 
+ 'Switzerland': 'CHE', 
+ 'Chile': 'CHL', 
+ 'China': 'CHN', 
+ 'Ivory Coast': 'CIV', 
+ 'Côte d\'Ivoire': 'CIV', 
+ 'Cameroon': 'CMR', 
+ 'Democratic Republic of the Congo': 'COD', 
+ 'Democratic Republic Of The Congo': 'COD', 
+ 'Democratic Republic of Congo': 'COD', 
+ 'Democratic Republic Of Congo': 'COD', 
+ 'Republic of the Congo': 'COG', 
+ 'Republic of Congo': 'COG', 
+ 'Congo': 'COG', 
+ 'Cook Islands': 'COK', 
+ 'Colombia': 'COL', 
+ 'Comoros': 'COM', 
+ 'Cape Verde': 'CPV', 
+ 'Cabo Verde': 'CPV', 
+ 'Costa Rica': 'CRI', 
+ 'Cuba': 'CUB', 
+ 'Curaçao': 'CUW', 
+ 'Christmas Island': 'CXR', 
+ 'Cayman Islands': 'CYM', 
+ 'Cyprus': 'CYP', 
+ 'Czech Republic': 'CZE', 
+ 'Czechia': 'CZE', 
+ 'Germany': 'DEU', 
+ 'Djibouti': 'DJI', 
+ 'Dominica': 'DMA', 
+ 'Denmark': 'DNK', 
+ 'Dominican Republic': 'DOM', 
+ 'Santo Domingo': 'DOM', 
+ 'Algeria': 'DZA', 
+ 'Ecuador': 'ECU', 
+ 'Egypt': 'EGY', 
+ 'Eritrea': 'ERI', 
+ 'Western Sahara': 'ESH', 
+ 'Spain': 'ESP', 
+ 'Estonia': 'EST', 
+ 'Ethiopia': 'ETH', 
+ 'Finland': 'FIN', 
+ 'Fiji': 'FJI', 
+ 'Falkland Islands': 'FLK', 
+ 'Falkland Islands (Malvinas)': 'FLK', 
+ 'France': 'FRA', 
+ 'Faroe Islands': 'FRO', 
+ 'Micronesia': 'FSM', 
+ 'Micronesia (Federated States of)': 'FSM', 
+ 'Gabon': 'GAB', 
+ 'United Kingdom': 'GBR', 
+ 'United Kingdom of Great Britain and Northern Ireland': 'GBR', 
+ 'England': 'GBR', 
+ 'Scotland': 'GBR', 
+ 'Wales': 'GBR', 
+ 'Northern Ireland': 'GBR', 
+ 'Georgia': 'GEO', 
+ 'Guernsey': 'GGY', 
+ 'Ghana': 'GHA', 
+ 'Gibraltar': 'GIB', 
+ 'Guinea': 'GIN', 
+ 'Guadeloupe': 'GLP', 
+ 'Gambia': 'GMB', 
+ 'Guinea-Bissau': 'GNB', 
+ 'Equatorial Guinea': 'GNQ', 
+ 'Greece': 'GRC', 
+ 'Grenada': 'GRD', 
+ 'Greenland': 'GRL', 
+ 'Guatemala': 'GTM', 
+ 'French Guiana': 'GUF', 
+ 'Guam': 'GUM', 
+ 'Guyana': 'GUY', 
+ 'Hong Kong': 'HKG', 
+ 'Heard Island and McDonald Islands': 'HMD', 
+ 'Heard Island And McDonald Islands': 'HMD', 
+ 'Honduras': 'HND', 
+ 'Croatia': 'HRV', 
+ 'Haiti': 'HTI', 
+ 'Hungary': 'HUN', 
+ 'Indonesia': 'IDN', 
+ 'Isle of Man': 'IMN', 
+ 'India': 'IND', 
+ 'British Indian Ocean Territory': 'IOT', 
+ 'Ireland': 'IRL', 
+ 'Iran': 'IRN', 
+ 'Iran (Islamic Republic of)': 'IRN', 
+ 'Iraq': 'IRQ', 
+ 'Iceland': 'ISL', 
+ 'Israel': 'ISR', 
+ 'Italy': 'ITA', 
+ 'Jamaica': 'JAM', 
+ 'Jersey': 'JEY', 
+ 'Jordan': 'JOR', 
+ 'Japan': 'JPN', 
+ 'Kazakhstan': 'KAZ', 
+ 'Kenya': 'KEN', 
+ 'Kyrgyzstan': 'KGZ', 
+ 'Cambodia': 'KHM', 
+ 'Kiribati': 'KIR', 
+ 'Saint Kitts and Nevis': 'KNA', 
+ 'Saint Kitts And Nevis': 'KNA', 
+ 'South Korea': 'KOR', 
+ 'Korea, Republic of': 'KOR', 
+ 'Kuwait': 'KWT', 
+ 'Laos': 'LAO', 
+ 'Lao People\'s Democratic Republic': 'LAO', 
+ 'Lebanon': 'LBN', 
+ 'Liberia': 'LBR', 
+ 'Libya': 'LBY', 
+ 'Saint Lucia': 'LCA', 
+ 'Liechtenstein': 'LIE', 
+ 'Sri Lanka': 'LKA', 
+ 'Lesotho': 'LSO', 
+ 'Lithuania': 'LTU', 
+ 'Luxembourg': 'LUX', 
+ 'Latvia': 'LVA', 
+ 'Macau': 'MAC', 
+ 'Macao': 'MAC', 
+ 'Saint Martin (French part)': 'MAF', 
+ 'Morocco': 'MAR', 
+ 'Monaco': 'MCO', 
+ 'Moldova': 'MDA', 
+ 'Moldova, Republic of': 'MDA', 
+ 'Madagascar': 'MDG', 
+ 'Maldives': 'MDV', 
+ 'Mexico': 'MEX', 
+ 'Marshall Islands': 'MHL', 
+ 'Macedonia': 'MKD', 
+ 'North Macedonia': 'MKD', 
+ 'Mali': 'MLI', 
+ 'Malta': 'MLT', 
+ 'Myanmar': 'MMR', 
+ 'Montenegro': 'MNE', 
+ 'Mongolia': 'MNG', 
+ 'Northern Mariana Islands': 'MNP', 
+ 'Mozambique': 'MOZ', 
+ 'Mauritania': 'MRT', 
+ 'Montserrat': 'MSR', 
+ 'Martinique': 'MTQ', 
+ 'Mauritius': 'MUS', 
+ 'Malawi': 'MWI', 
+ 'Malaysia': 'MYS', 
+ 'Mayotte': 'MYT', 
+ 'Namibia': 'NAM', 
+ 'New Caledonia': 'NCL', 
+ 'Niger': 'NER', 
+ 'Norfolk Island': 'NFK', 
+ 'Nigeria': 'NGA', 
+ 'Nicaragua': 'NIC', 
+ 'Niue': 'NIU', 
+ 'Netherlands': 'NLD', 
+ 'Norway': 'NOR', 
+ 'Nepal': 'NPL', 
+ 'Nauru': 'NRU', 
+ 'New Zealand': 'NZL', 
+ 'Oman': 'OMN', 
+ 'Pakistan': 'PAK', 
+ 'Panama': 'PAN', 
+ 'Pitcairn Islands': 'PCN', 
+ 'Pitcairn': 'PCN', 
+ 'Peru': 'PER', 
+ 'Philippines': 'PHL', 
+ 'Palau': 'PLW', 
+ 'Papua New Guinea': 'PNG', 
+ 'Poland': 'POL', 
+ 'Puerto Rico': 'PRI', 
+ 'North Korea': 'PRK', 
+ 'Korea (Democratic People\'s Republic of)': 'PRK', 
+ 'Portugal': 'PRT', 
+ 'Paraguay': 'PRY', 
+ 'Palestine': 'PSE', 
+ 'Palestine, State of': 'PSE', 
+ 'French Polynesia': 'PYF', 
+ 'Qatar': 'QAT', 
+ 'Réunion': 'REU', 
+ 'Kosovo': 'RKS', 
+ 'Romania': 'ROU', 
+ 'Russia': 'RUS', 
+ 'Russian Federation': 'RUS', 
+ 'Rwanda': 'RWA', 
+ 'Saudi Arabia': 'SAU', 
+ 'Sudan': 'SDN', 
+ 'Senegal': 'SEN', 
+ 'Singapore': 'SGP', 
+ 'South Georgia and the South Sandwich Islands': 'SGS', 
+ 'South Georgia And The South Sandwich Islands': 'SGS', 
+ 'South Georgia and South Sandwich Islands': 'SGS', 
+ 'South Georgia And South Sandwich Islands': 'SGS', 
+ 'Saint Helena': 'SHN', 
+ 'Saint Helena, Ascension and Tristan da Cunha': 'SHN', 
+ 'Svalbard and Jan Mayen': 'SJM', 
+ 'Svalbard And Jan Mayen': 'SJM', 
+ 'Solomon Islands': 'SLB', 
+ 'Sierra Leone': 'SLE', 
+ 'El Salvador': 'SLV', 
+ 'San Marino': 'SMR', 
+ 'Somalia': 'SOM', 
+ 'Saint Pierre and Miquelon': 'SPM', 
+ 'Serbia': 'SRB', 
+ 'South Sudan': 'SSD', 
+ 'São Tomé and Príncipe': 'STP', 
+ 'Sao Tome and Principe': 'STP', 
+ 'São Tomé And Príncipe': 'STP', 
+ 'Suriname': 'SUR', 
+ 'Slovakia': 'SVK', 
+ 'Slovenia': 'SVN', 
+ 'Sweden': 'SWE', 
+ 'Swaziland': 'SWZ', 
+ 'Eswatini': 'SWZ', 
+ 'Sint Maarten': 'SXM', 
+ 'Sint Maarten (Dutch part)': 'SXM', 
+ 'Seychelles': 'SYC', 
+ 'Syria': 'SYR', 
+ 'Syrian Arab Republic': 'SYR', 
+ 'Turks and Caicos Islands': 'TCA', 
+ 'Turks And Caicos Islands': 'TCA', 
+ 'Chad': 'TCD', 
+ 'Togo': 'TGO', 
+ 'Thailand': 'THA', 
+ 'Tajikistan': 'TJK', 
+ 'Tokelau': 'TKL', 
+ 'Turkmenistan': 'TKM', 
+ 'East Timor': 'TLS', 
+ 'Timor-Leste': 'TLS', 
+ 'Tonga': 'TON', 
+ 'Trinidad and Tobago': 'TTO', 
+ 'Tunisia': 'TUN', 
+ 'Turkey': 'TUR', 
+ 'Türkiye': 'TUR', 
+ 'Tuvalu': 'TUV', 
+ 'Taiwan': 'TWN', 
+ 'Taiwan, Province of China': 'TWN', 
+ 'Tanzania': 'TZA', 
+ 'Tanzania, United Republic of': 'TZA', 
+ 'Uganda': 'UGA', 
+ 'Ukraine': 'UKR', 
+ 'United States Minor Outlying Islands': 'UMI', 
+ 'Uruguay': 'URY', 
+ 'United States of America': 'USA', 
+ 'United States': 'USA', 
+ 'Uzbekistan': 'UZB', 
+ 'Vatican City': 'VAT', 
+ 'Holy See': 'VAT', 
+ 'Saint Vincent and the Grenadines': 'VCT', 
+ 'Saint Vincent And the Grenadines': 'VCT', 
+ 'Venezuela': 'VEN', 
+ 'Venezuela (Bolivarian Republic of)': 'VEN', 
+ 'British Virgin Islands': 'VGB', 
+ 'Virgin Islands (British)': 'VGB', 
+ 'United States Virgin Islands': 'VIR', 
+ 'Virgin Islands (U.S.)': 'VIR', 
+ 'Vietnam': 'VNM', 
+ 'Viet Nam': 'VNM', 
+ 'Viet-Nam': 'VNM', 
+ 'Vanuatu': 'VUT', 
+ 'Wallis and Futuna': 'WLF', 
+ 'Samoa': 'WSM', 
+ 'Yemen': 'YEM', 
+ 'South Africa': 'ZAF', 
+ 'Zambia': 'ZMB', 
+ 'Zimbabwe': 'ZWE'
+ }
+
+all_countries =['Afghanistan',
+ 'Åland Islands',
+ 'Albania',
+ 'Algeria',
+ 'American Samoa',
+ 'Andorra',
+ 'Angola',
+ 'Anguilla',
+ 'Antarctica',
+ 'Antigua and Barbuda',
+ 'Argentina',
+ 'Armenia',
+ 'Aruba',
+ 'Australia',
+ 'Austria',
+ 'Azerbaijan',
+ 'Bahamas',
+ 'Bahrain',
+ 'Bangladesh',
+ 'Barbados',
+ 'Belarus',
+ 'Belgium',
+ 'Belize',
+ 'Benin',
+ 'Bermuda',
+ 'Bhutan',
+ 'Bolivia',
+ 'Bosnia and Herzegovina',
+ 'Botswana',
+ 'Bouvet Island',
+ 'Brazil',
+ 'British Indian Ocean Territory',
+ 'British Virgin Islands',
+ 'Brunei',
+ 'Bulgaria',
+ 'Burkina Faso',
+ 'Burundi',
+ 'Cambodia',
+ 'Cameroon',
+ 'Canada',
+ 'Cape Verde',
+ 'Caribbean Netherlands',
+ 'Cayman Islands',
+ 'Central African Republic',
+ 'Chad',
+ 'Chile',
+ 'China',
+ 'Christmas Island',
+ 'Cocos Islands',
+ 'Colombia',
+ 'Comoros',
+ 'Cook Islands',
+ 'Costa Rica',
+ 'Croatia',
+ 'Cuba',
+ 'Curaçao',
+ 'Cyprus',
+ 'Czech Republic',
+ 'Democratic Republic of the Congo',
+ 'Denmark',
+ 'Djibouti',
+ 'Dominica',
+ 'Dominican Republic',
+ 'East Timor',
+ 'Ecuador',
+ 'Egypt',
+ 'El Salvador',
+ 'Equatorial Guinea',
+ 'Eritrea',
+ 'Estonia',
+ 'Ethiopia',
+ 'Falkland Islands',
+ 'Faroe Islands',
+ 'Fiji',
+ 'Finland',
+ 'France',
+ 'French Guiana',
+ 'French Polynesia',
+ 'French Southern Territories',
+ 'Gabon',
+ 'Gambia',
+ 'Georgia',
+ 'Germany',
+ 'Ghana',
+ 'Gibraltar',
+ 'Greece',
+ 'Greenland',
+ 'Grenada',
+ 'Guadeloupe',
+ 'Guam',
+ 'Guatemala',
+ 'Guernsey',
+ 'Guinea',
+ 'Guinea-Bissau',
+ 'Guyana',
+ 'Haiti',
+ 'Heard Island and McDonald Islands',
+ 'Honduras',
+ 'Hong Kong',
+ 'Hungary',
+ 'Iceland',
+ 'India',
+ 'Indonesia',
+ 'Iran',
+ 'Iraq',
+ 'Ireland',
+ 'Isle of Man',
+ 'Israel',
+ 'Italy',
+ 'Ivory Coast',
+ 'Jamaica',
+ 'Japan',
+ 'Jersey',
+ 'Jordan',
+ 'Kazakhstan',
+ 'Kenya',
+ 'Kiribati',
+ 'Kosovo',
+ 'Kuwait',
+ 'Kyrgyzstan',
+ 'Laos',
+ 'Latvia',
+ 'Lebanon',
+ 'Lesotho',
+ 'Liberia',
+ 'Libya',
+ 'Liechtenstein',
+ 'Lithuania',
+ 'Luxembourg',
+ 'Macau',
+ 'Macedonia',
+ 'Madagascar',
+ 'Malawi',
+ 'Malaysia',
+ 'Maldives',
+ 'Mali',
+ 'Malta',
+ 'Marshall Islands',
+ 'Martinique',
+ 'Mauritania',
+ 'Mauritius',
+ 'Mayotte',
+ 'Mexico',
+ 'Micronesia',
+ 'Moldova',
+ 'Monaco',
+ 'Mongolia',
+ 'Montenegro',
+ 'Montserrat',
+ 'Morocco',
+ 'Mozambique',
+ 'Myanmar',
+ 'Namibia',
+ 'Nauru',
+ 'Nepal',
+ 'Netherlands',
+ 'New Caledonia',
+ 'New Zealand',
+ 'Nicaragua',
+ 'Niger',
+ 'Nigeria',
+ 'Niue',
+ 'Norfolk Island',
+ 'North Korea',
+ 'Northern Mariana Islands',
+ 'Norway',
+ 'Oman',
+ 'Pakistan',
+ 'Palau',
+ 'Palestine',
+ 'Panama',
+ 'Papua New Guinea',
+ 'Paraguay',
+ 'Peru',
+ 'Philippines',
+ 'Pitcairn Islands',
+ 'Poland',
+ 'Portugal',
+ 'Puerto Rico',
+ 'Qatar',
+ 'Republic of the Congo',
+ 'Romania',
+ 'Russia',
+ 'Rwanda',
+ 'Réunion',
+ 'Saint Barthélemy',
+ 'Saint Helena',
+ 'Saint Kitts and Nevis',
+ 'Saint Lucia',
+ 'Saint Martin (French part)',
+ 'Saint Pierre and Miquelon',
+ 'Saint Vincent and the Grenadines',
+ 'Samoa',
+ 'San Marino',
+ 'Saudi Arabia',
+ 'Senegal',
+ 'Serbia',
+ 'Seychelles',
+ 'Sierra Leone',
+ 'Singapore',
+ 'Sint Maarten',
+ 'Slovakia',
+ 'Slovenia',
+ 'Solomon Islands',
+ 'Somalia',
+ 'South Africa',
+ 'South Georgia and the South Sandwich Islands',
+ 'South Korea',
+ 'South Sudan',
+ 'Spain',
+ 'Sri Lanka',
+ 'Sudan',
+ 'Suriname',
+ 'Svalbard and Jan Mayen',
+ 'Swaziland',
+ 'Sweden',
+ 'Switzerland',
+ 'Syria',
+ 'São Tomé and Príncipe',
+ 'Taiwan',
+ 'Tajikistan',
+ 'Tanzania',
+ 'Thailand',
+ 'Togo',
+ 'Tokelau',
+ 'Tonga',
+ 'Trinidad and Tobago',
+ 'Tunisia',
+ 'Turkey',
+ 'Turkmenistan',
+ 'Turks and Caicos Islands',
+ 'Tuvalu',
+ 'Uganda',
+ 'Ukraine',
+ 'United Arab Emirates',
+ 'United Kingdom',
+ 'United States Minor Outlying Islands',
+ 'United States Virgin Islands',
+ 'United States of America',
+ 'Uruguay',
+ 'Uzbekistan',
+ 'Vanuatu',
+ 'Vatican City',
+ 'Venezuela',
+ 'Vietnam',
+ 'Wallis and Futuna',
+ 'Western Sahara',
+ 'Yemen',
+ 'Zambia',
+ 'Zimbabwe'
+ ]
+
+iso3_to_country= {
+ 'ABW': 'Aruba', 
+ 'AFG': 'Afghanistan', 
+ 'AGO': 'Angola', 
+ 'AIA': 'Anguilla', 
+ 'ALA': 'Åland Islands', 
+ 'ALB': 'Albania', 
+ 'AND': 'Andorra', 
+ 'ARE': 'United Arab Emirates', 
+ 'ARG': 'Argentina', 
+ 'ARM': 'Armenia', 
+ 'ASM': 'American Samoa', 
+ 'ATA': 'Antarctica', 
+ 'ATF': 'French Southern Territories', 
+ 'ATG': 'Antigua and Barbuda', 
+ 'AUS': 'Australia', 
+ 'AUT': 'Austria', 
+ 'AZE': 'Azerbaijan', 
+ 'BDI': 'Burundi', 
+ 'BEL': 'Belgium', 
+ 'BEN': 'Benin', 
+ 'BES': 'Caribbean Netherlands', 
+ 'BFA': 'Burkina Faso', 
+ 'BGD': 'Bangladesh', 
+ 'BGR': 'Bulgaria', 
+ 'BHR': 'Bahrain', 
+ 'BHS': 'Bahamas', 
+ 'BIH': 'Bosnia and Herzegovina', 
+ 'BLM': 'Saint Barthélemy', 
+ 'BLR': 'Belarus', 
+ 'BLZ': 'Belize', 
+ 'BMU': 'Bermuda', 
+ 'BOL': 'Bolivia', 
+ 'BRA': 'Brazil', 
+ 'BRB': 'Barbados', 
+ 'BRN': 'Brunei', 
+ 'BTN': 'Bhutan', 
+ 'BVT': 'Bouvet Island', 
+ 'BWA': 'Botswana', 
+ 'CAF': 'Central African Republic', 
+ 'CAN': 'Canada', 
+ 'CCK': 'Cocos Islands', 
+ 'CHE': 'Switzerland', 
+ 'CHL': 'Chile', 
+ 'CHN': 'China', 
+ 'CIV': 'Ivory Coast', 
+ 'CMR': 'Cameroon', 
+ 'COD': 'Democratic Republic of the Congo', 
+ 'COG': 'Republic of the Congo', 
+ 'COK': 'Cook Islands', 
+ 'COL': 'Colombia', 
+ 'COM': 'Comoros', 
+ 'CPV': 'Cape Verde', 
+ 'CRI': 'Costa Rica', 
+ 'CUB': 'Cuba', 
+ 'CUW': 'Curaçao', 
+ 'CXR': 'Christmas Island', 
+ 'CYM': 'Cayman Islands', 
+ 'CYP': 'Cyprus', 
+ 'CZE': 'Czech Republic', 
+ 'DEU': 'Germany', 
+ 'DJI': 'Djibouti', 
+ 'DMA': 'Dominica', 
+ 'DNK': 'Denmark', 
+ 'DOM': 'Dominican Republic', 
+ 'DZA': 'Algeria', 
+ 'ECU': 'Ecuador', 
+ 'EGY': 'Egypt', 
+ 'ERI': 'Eritrea', 
+ 'ESH': 'Western Sahara', 
+ 'ESP': 'Spain', 
+ 'EST': 'Estonia', 
+ 'ETH': 'Ethiopia', 
+ 'FIN': 'Finland', 
+ 'FJI': 'Fiji', 
+ 'FLK': 'Falkland Islands', 
+ 'FRA': 'France', 
+ 'FRO': 'Faroe Islands', 
+ 'FSM': 'Micronesia', 
+ 'GAB': 'Gabon', 
+ 'GBR': 'England', 
+ 'GEO': 'Georgia', 
+ 'GGY': 'Guernsey', 
+ 'GHA': 'Ghana', 
+ 'GIB': 'Gibraltar', 
+ 'GIN': 'Guinea', 
+ 'GLP': 'Guadeloupe', 
+ 'GMB': 'Gambia', 
+ 'GNB': 'Guinea-Bissau', 
+ 'GNQ': 'Equatorial Guinea', 
+ 'GRC': 'Greece', 
+ 'GRD': 'Grenada', 
+ 'GRL': 'Greenland', 
+ 'GTM': 'Guatemala', 
+ 'GUF': 'French Guiana', 
+ 'GUM': 'Guam', 
+ 'GUY': 'Guyana', 
+ 'HKG': 'Hong Kong', 
+ 'HMD': 'Heard Island and McDonald Islands', 
+ 'HND': 'Honduras', 
+ 'HRV': 'Croatia', 
+ 'HTI': 'Haiti', 
+ 'HUN': 'Hungary', 
+ 'IDN': 'Indonesia', 
+ 'IMN': 'Isle of Man', 
+ 'IND': 'India', 
+ 'IOT': 'British Indian Ocean Territory', 
+ 'IRL': 'Ireland', 
+ 'IRN': 'Iran', 
+ 'IRQ': 'Iraq', 
+ 'ISL': 'Iceland', 
+ 'ISR': 'Israel', 
+ 'ITA': 'Italy', 
+ 'JAM': 'Jamaica', 
+ 'JEY': 'Jersey', 
+ 'JOR': 'Jordan', 
+ 'JPN': 'Japan', 
+ 'KAZ': 'Kazakhstan', 
+ 'KEN': 'Kenya', 
+ 'KGZ': 'Kyrgyzstan', 
+ 'KHM': 'Cambodia', 
+ 'KIR': 'Kiribati', 
+ 'KNA': 'Saint Kitts and Nevis', 
+ 'KOR': 'South Korea', 
+ 'KWT': 'Kuwait', 
+ 'LAO': 'Laos', 
+ 'LBN': 'Lebanon', 
+ 'LBR': 'Liberia', 
+ 'LBY': 'Libya', 
+ 'LCA': 'Saint Lucia', 
+ 'LIE': 'Liechtenstein', 
+ 'LKA': 'Sri Lanka', 
+ 'LSO': 'Lesotho', 
+ 'LTU': 'Lithuania', 
+ 'LUX': 'Luxembourg', 
+ 'LVA': 'Latvia', 
+ 'MAC': 'Macau', 
+ 'MAF': 'Saint Martin (French part)', 
+ 'MAR': 'Morocco', 
+ 'MCO': 'Monaco', 
+ 'MDA': 'Moldova', 
+ 'MDG': 'Madagascar', 
+ 'MDV': 'Maldives', 
+ 'MEX': 'Mexico', 
+ 'MHL': 'Marshall Islands', 
+ 'MKD': 'Macedonia', 
+ 'MLI': 'Mali', 
+ 'MLT': 'Malta', 
+ 'MMR': 'Myanmar', 
+ 'MNE': 'Montenegro', 
+ 'MNG': 'Mongolia', 
+ 'MNP': 'Northern Mariana Islands', 
+ 'MOZ': 'Mozambique', 
+ 'MRT': 'Mauritania', 
+ 'MSR': 'Montserrat', 
+ 'MTQ': 'Martinique', 
+ 'MUS': 'Mauritius', 
+ 'MWI': 'Malawi', 
+ 'MYS': 'Malaysia', 
+ 'MYT': 'Mayotte', 
+ 'NAM': 'Namibia', 
+ 'NCL': 'New Caledonia', 
+ 'NER': 'Niger', 
+ 'NFK': 'Norfolk Island', 
+ 'NGA': 'Nigeria', 
+ 'NIC': 'Nicaragua', 
+ 'NIU': 'Niue', 
+ 'NLD': 'Netherlands', 
+ 'NOR': 'Norway', 
+ 'NPL': 'Nepal', 
+ 'NRU': 'Nauru', 
+ 'NZL': 'New Zealand', 
+ 'OMN': 'Oman', 
+ 'PAK': 'Pakistan', 
+ 'PAN': 'Panama', 
+ 'PCN': 'Pitcairn Islands', 
+ 'PER': 'Peru', 
+ 'PHL': 'Philippines', 
+ 'PLW': 'Palau', 
+ 'PNG': 'Papua New Guinea', 
+ 'POL': 'Poland', 
+ 'PRI': 'Puerto Rico', 
+ 'PRK': 'North Korea', 
+ 'PRT': 'Portugal', 
+ 'PRY': 'Paraguay', 
+ 'PSE': 'Palestine', 
+ 'PYF': 'French Polynesia', 
+ 'QAT': 'Qatar', 
+ 'REU': 'Réunion', 
+ 'RKS': 'Kosovo', 
+ 'ROU': 'Romania', 
+ 'RUS': 'Russia', 
+ 'RWA': 'Rwanda', 
+ 'SAU': 'Saudi Arabia', 
+ 'SDN': 'Sudan', 
+ 'SEN': 'Senegal', 
+ 'SGP': 'Singapore', 
+ 'SGS': 'South Georgia and the South Sandwich Islands', 
+ 'SHN': 'Saint Helena', 
+ 'SJM': 'Svalbard and Jan Mayen', 
+ 'SLB': 'Solomon Islands', 
+ 'SLE': 'Sierra Leone', 
+ 'SLV': 'El Salvador', 
+ 'SMR': 'San Marino', 
+ 'SOM': 'Somalia', 
+ 'SPM': 'Saint Pierre and Miquelon', 
+ 'SRB': 'Serbia', 
+ 'SSD': 'South Sudan', 
+ 'STP': 'São Tomé and Príncipe', 
+ 'SUR': 'Suriname', 
+ 'SVK': 'Slovakia', 
+ 'SVN': 'Slovenia', 
+ 'SWE': 'Sweden', 
+ 'SWZ': 'Swaziland', 
+ 'SXM': 'Sint Maarten', 
+ 'SYC': 'Seychelles', 
+ 'SYR': 'Syria', 
+ 'TCA': 'Turks and Caicos Islands', 
+ 'TCD': 'Chad', 
+ 'TGO': 'Togo', 
+ 'THA': 'Thailand', 
+ 'TJK': 'Tajikistan', 
+ 'TKL': 'Tokelau', 
+ 'TKM': 'Turkmenistan', 
+ 'TLS': 'East Timor', 
+ 'TON': 'Tonga', 
+ 'TTO': 'Trinidad and Tobago', 
+ 'TUN': 'Tunisia', 
+ 'TUR': 'Turkey', 
+ 'TUV': 'Tuvalu', 
+ 'TWN': 'Taiwan', 
+ 'TZA': 'Tanzania', 
+ 'UGA': 'Uganda', 
+ 'UKR': 'Ukraine', 
+ 'UMI': 'United States Minor Outlying Islands', 
+ 'URY': 'Uruguay', 
+ 'USA': 'United States of America', 
+ 'UZB': 'Uzbekistan', 
+ 'VAT': 'Vatican City', 
+ 'VCT': 'Saint Vincent and the Grenadines', 
+ 'VEN': 'Venezuela', 
+ 'VGB': 'British Virgin Islands', 
+ 'VIR': 'United States Virgin Islands', 
+ 'VNM': 'Vietnam', 
+ 'VUT': 'Vanuatu', 
+ 'WLF': 'Wallis and Futuna', 
+ 'WSM': 'Samoa', 
+ 'YEM': 'Yemen', 
+ 'ZAF': 'South Africa', 
+ 'ZMB': 'Zambia', 
+ 'ZWE': 'Zimbabwe'
+ }
+
+iso2_to_iso3 = {
+ 'AD': 'AND', 
+ 'AE': 'ARE', 
+ 'AF': 'AFG', 
+ 'AG': 'ATG', 
+ 'AI': 'AIA', 
+ 'AL': 'ALB', 
+ 'AM': 'ARM', 
+ 'AO': 'AGO', 
+ 'AQ': 'ATA', 
+ 'AR': 'ARG', 
+ 'AS': 'ASM', 
+ 'AT': 'AUT', 
+ 'AU': 'AUS', 
+ 'AW': 'ABW', 
+ 'AX': 'ALA', 
+ 'AZ': 'AZE', 
+ 'BA': 'BIH', 
+ 'BB': 'BRB', 
+ 'BD': 'BGD', 
+ 'BE': 'BEL', 
+ 'BF': 'BFA', 
+ 'BG': 'BGR', 
+ 'BH': 'BHR', 
+ 'BI': 'BDI', 
+ 'BJ': 'BEN', 
+ 'BL': 'BLM', 
+ 'BM': 'BMU', 
+ 'BN': 'BRN', 
+ 'BO': 'BOL', 
+ 'BQ': 'BES', 
+ 'BR': 'BRA', 
+ 'BS': 'BHS', 
+ 'BT': 'BTN', 
+ 'BV': 'BVT', 
+ 'BW': 'BWA', 
+ 'BY': 'BLR', 
+ 'BZ': 'BLZ', 
+ 'CA': 'CAN', 
+ 'CC': 'CCK', 
+ 'CD': 'COD', 
+ 'CF': 'CAF', 
+ 'CG': 'COG', 
+ 'CH': 'CHE', 
+ 'CI': 'CIV', 
+ 'CK': 'COK', 
+ 'CL': 'CHL', 
+ 'CM': 'CMR', 
+ 'CN': 'CHN', 
+ 'CO': 'COL', 
+ 'CR': 'CRI', 
+ 'CU': 'CUB', 
+ 'CV': 'CPV', 
+ 'CW': 'CUW', 
+ 'CX': 'CXR', 
+ 'CY': 'CYP', 
+ 'CZ': 'CZE', 
+ 'DE': 'DEU', 
+ 'DJ': 'DJI', 
+ 'DK': 'DNK', 
+ 'DM': 'DMA', 
+ 'DO': 'DOM', 
+ 'DZ': 'DZA', 
+ 'EC': 'ECU', 
+ 'EE': 'EST', 
+ 'EG': 'EGY', 
+ 'EH': 'ESH', 
+ 'ER': 'ERI', 
+ 'ES': 'ESP', 
+ 'ET': 'ETH', 
+ 'FI': 'FIN', 
+ 'FJ': 'FJI', 
+ 'FK': 'FLK', 
+ 'FM': 'FSM', 
+ 'FO': 'FRO', 
+ 'FR': 'FRA', 
+ 'GA': 'GAB', 
+ 'GB': 'GBR', 
+ 'GD': 'GRD', 
+ 'GE': 'GEO', 
+ 'GF': 'GUF', 
+ 'GG': 'GGY', 
+ 'GH': 'GHA', 
+ 'GI': 'GIB', 
+ 'GL': 'GRL', 
+ 'GM': 'GMB', 
+ 'GN': 'GIN', 
+ 'GP': 'GLP', 
+ 'GQ': 'GNQ', 
+ 'GR': 'GRC', 
+ 'GS': 'SGS', 
+ 'GT': 'GTM', 
+ 'GU': 'GUM', 
+ 'GW': 'GNB', 
+ 'GY': 'GUY', 
+ 'HK': 'HKG', 
+ 'HM': 'HMD', 
+ 'HN': 'HND', 
+ 'HR': 'HRV', 
+ 'HT': 'HTI', 
+ 'HU': 'HUN', 
+ 'ID': 'IDN', 
+ 'IE': 'IRL', 
+ 'IL': 'ISR', 
+ 'IM': 'IMN', 
+ 'IN': 'IND', 
+ 'IO': 'IOT', 
+ 'IQ': 'IRQ', 
+ 'IR': 'IRN', 
+ 'IS': 'ISL', 
+ 'IT': 'ITA', 
+ 'JE': 'JEY', 
+ 'JM': 'JAM', 
+ 'JO': 'JOR', 
+ 'JP': 'JPN', 
+ 'KE': 'KEN', 
+ 'KG': 'KGZ', 
+ 'KH': 'KHM', 
+ 'KI': 'KIR', 
+ 'KM': 'COM', 
+ 'KN': 'KNA', 
+ 'KP': 'PRK', 
+ 'KR': 'KOR', 
+ 'KW': 'KWT', 
+ 'KY': 'CYM', 
+ 'KZ': 'KAZ', 
+ 'LA': 'LAO', 
+ 'LB': 'LBN', 
+ 'LC': 'LCA', 
+ 'LI': 'LIE', 
+ 'LK': 'LKA', 
+ 'LR': 'LBR', 
+ 'LS': 'LSO', 
+ 'LT': 'LTU', 
+ 'LU': 'LUX', 
+ 'LV': 'LVA', 
+ 'LY': 'LBY', 
+ 'MA': 'MAR', 
+ 'MC': 'MCO', 
+ 'MD': 'MDA', 
+ 'ME': 'MNE', 
+ 'MF': 'MAF', 
+ 'MG': 'MDG', 
+ 'MH': 'MHL', 
+ 'MK': 'MKD', 
+ 'ML': 'MLI', 
+ 'MM': 'MMR', 
+ 'MN': 'MNG', 
+ 'MO': 'MAC', 
+ 'MP': 'MNP', 
+ 'MQ': 'MTQ', 
+ 'MR': 'MRT', 
+ 'MS': 'MSR', 
+ 'MT': 'MLT', 
+ 'MU': 'MUS', 
+ 'MV': 'MDV', 
+ 'MW': 'MWI', 
+ 'MX': 'MEX', 
+ 'MY': 'MYS', 
+ 'MZ': 'MOZ', 
+ 'NA': 'NAM', 
+ 'NC': 'NCL', 
+ 'NE': 'NER', 
+ 'NF': 'NFK', 
+ 'NG': 'NGA', 
+ 'NI': 'NIC', 
+ 'NL': 'NLD', 
+ 'NO': 'NOR', 
+ 'NP': 'NPL', 
+ 'NR': 'NRU', 
+ 'NU': 'NIU', 
+ 'NZ': 'NZL', 
+ 'OM': 'OMN', 
+ 'PA': 'PAN', 
+ 'PE': 'PER', 
+ 'PF': 'PYF', 
+ 'PG': 'PNG', 
+ 'PH': 'PHL', 
+ 'PK': 'PAK', 
+ 'PL': 'POL', 
+ 'PM': 'SPM', 
+ 'PN': 'PCN', 
+ 'PR': 'PRI', 
+ 'PS': 'PSE', 
+ 'PT': 'PRT', 
+ 'PW': 'PLW', 
+ 'PY': 'PRY', 
+ 'QA': 'QAT', 
+ 'RE': 'REU', 
+ 'RO': 'ROU', 
+ 'RS': 'SRB', 
+ 'RU': 'RUS', 
+ 'RW': 'RWA', 
+ 'SA': 'SAU', 
+ 'SB': 'SLB', 
+ 'SC': 'SYC', 
+ 'SD': 'SDN', 
+ 'SE': 'SWE', 
+ 'SG': 'SGP', 
+ 'SH': 'SHN', 
+ 'SI': 'SVN', 
+ 'SJ': 'SJM', 
+ 'SK': 'SVK', 
+ 'SL': 'SLE', 
+ 'SM': 'SMR', 
+ 'SN': 'SEN', 
+ 'SO': 'SOM', 
+ 'SR': 'SUR', 
+ 'SS': 'SSD', 
+ 'ST': 'STP', 
+ 'SV': 'SLV', 
+ 'SX': 'SXM', 
+ 'SY': 'SYR', 
+ 'SZ': 'SWZ', 
+ 'TC': 'TCA', 
+ 'TD': 'TCD', 
+ 'TF': 'ATF', 
+ 'TG': 'TGO', 
+ 'TH': 'THA', 
+ 'TJ': 'TJK', 
+ 'TK': 'TKL', 
+ 'TL': 'TLS', 
+ 'TM': 'TKM', 
+ 'TN': 'TUN', 
+ 'TO': 'TON', 
+ 'TR': 'TUR', 
+ 'TT': 'TTO', 
+ 'TV': 'TUV', 
+ 'TW': 'TWN', 
+ 'TZ': 'TZA', 
+ 'UA': 'UKR', 
+ 'UG': 'UGA', 
+ 'UM': 'UMI', 
+ 'US': 'USA', 
+ 'UY': 'URY', 
+ 'UZ': 'UZB', 
+ 'VA': 'VAT', 
+ 'VC': 'VCT', 
+ 'VE': 'VEN', 
+ 'VG': 'VGB', 
+ 'VI': 'VIR', 
+ 'VN': 'VNM', 
+ 'VU': 'VUT', 
+ 'WF': 'WLF', 
+ 'WS': 'WSM', 
+ 'XK': 'RKS', 
+ 'YE': 'YEM', 
+ 'YT': 'MYT', 
+ 'ZA': 'ZAF', 
+ 'ZM': 'ZMB', 
+ 'ZW': 'ZWE'
+ }
+
+all_iso2 = list(iso2_to_iso3.keys()) 
+all_iso3 = list(iso2_to_iso3.values())
+iso3_to_iso2 = rev_dict(iso2_to_iso3)
+iso2_to_country = {k: iso3_to_country[v] for k,v in iso2_to_iso3.items()}
+country_to_iso2 = rev_dict(iso2_to_country)
+
+country_to_location_key = {'Andorra': 'AD',
+ 'United Arab Emirates': 'AE',
+ 'Afghanistan': 'AF',
+ 'Antigua and Barbuda': 'AG',
+ 'Antigua And Barbuda': 'AG',
+ 'Anguilla': 'AI',
+ 'Åland Islands': None,
+ 'Albania': 'AL',
+ 'Armenia': 'AM',
+ 'Angola': 'AO',
+ 'Antarctica': 'AQ',
+ 'Argentina': 'AR',
+ 'American Samoa': 'AS',
+ 'Austria': 'AT',
+ 'Australia': 'AU',
+ 'Aruba': 'AW',
+ 'Azerbaijan': 'AZ',
+ 'Bosnia and Herzegovina': 'BA',
+ 'Bosnia And Herzegovina': 'BA',
+ 'Barbados': 'BB',
+ 'Bangladesh': 'BD',
+ 'Belgium': 'BE',
+ 'Burkina Faso': 'BF',
+ 'Bulgaria': 'BG',
+ 'Bahrain': 'BH',
+ 'Burundi': 'BI',
+ 'Benin': 'BJ',
+ 'Saint Barthélemy': 'FR_BL',
+ 'Bermuda': 'BM',
+ 'Brunei': 'BN',
+ 'Brunei Darussalam': 'BN',
+ 'Bolivia': 'BO',
+ 'Bolivia (Plurinational State of)': 'BO',
+ 'Caribbean Netherlands': 'BQ',
+ 'Bonaire, Sint Eustatius and Saba': 'BQ',
+ 'Brazil': 'BR',
+ 'Bahamas': 'BS',
+ 'Bhutan': 'BT',
+ 'Bouvet Island': 'BV',
+ 'Botswana': 'BW',
+ 'Belarus': 'BY',
+ 'Belize': 'BZ',
+ 'Canada': 'CA',
+ 'Cocos Islands': 'CC',
+ 'Cocos (Keeling) Islands': 'CC',
+ 'Democratic Republic of the Congo': 'CD',
+ 'Democratic Republic Of The Congo': 'CD',
+ 'Democratic Republic of Congo': 'CD',
+ 'Democratic Republic Of Congo': 'CD',
+ 'Central African Republic': 'CF',
+ 'Republic of the Congo': 'CG',
+ 'Republic of Congo': 'CG',
+ 'Congo': 'CG',
+ 'Switzerland': 'CH',
+ 'Ivory Coast': 'CI',
+ "Côte d'Ivoire": 'CI',
+ 'Cook Islands': 'CK',
+ 'Chile': 'CL',
+ 'Cameroon': 'CM',
+ 'China': 'CN',
+ 'Colombia': 'CO',
+ 'Costa Rica': 'CR',
+ 'Cuba': 'CU',
+ 'Cape Verde': 'CV',
+ 'Cabo Verde': 'CV',
+ 'Curaçao': 'CW',
+ 'Christmas Island': 'CX',
+ 'Cyprus': 'CY',
+ 'Czech Republic': 'CZ',
+ 'Czechia': 'CZ',
+ 'Germany': 'DE',
+ 'Djibouti': 'DJ',
+ 'Denmark': 'DK',
+ 'Dominica': 'DM',
+ 'Dominican Republic': 'DO',
+ 'Santo Domingo': 'DO',
+ 'Algeria': 'DZ',
+ 'Ecuador': 'EC',
+ 'Estonia': 'EE',
+ 'Egypt': 'EG',
+ 'Western Sahara': 'EH',
+ 'Eritrea': 'ER',
+ 'Spain': 'ES',
+ 'Ethiopia': 'ET',
+ 'Finland': 'FI',
+ 'Fiji': 'FJ',
+ 'Falkland Islands': 'FK',
+ 'Falkland Islands (Malvinas)': 'FK',
+ 'Micronesia': 'FM',
+ 'Micronesia (Federated States of)': 'FM',
+ 'Faroe Islands': 'FO',
+ 'France': 'FR',
+ 'Gabon': 'GA',
+ 'United Kingdom': 'GB',
+ 'United Kingdom of Great Britain and Northern Ireland': 'GB',
+ 'England': 'GB_ENG',
+ 'Scotland': 'GB_SCT',
+ 'Wales': 'GB_WLS',
+ 'Northern Ireland': 'GB_NIR',
+ 'Grenada': 'GD',
+ 'Georgia': 'GE',
+ 'French Guiana': 'GF',
+ 'Guernsey': 'GG',
+ 'Ghana': 'GH',
+ 'Gibraltar': 'GI',
+ 'Guadeloupe': 'FR_GUA',
+ 'Greenland': 'GL',
+ 'Gambia': 'GM',
+ 'Guinea': 'GN',
+ 'Equatorial Guinea': 'GQ',
+ 'Greece': 'GR',
+ 'South Georgia and the South Sandwich Islands': 'GS',
+ 'South Georgia And The South Sandwich Islands': 'GS',
+ 'South Georgia and South Sandwich Islands': 'GS',
+ 'South Georgia And South Sandwich Islands': 'GS',
+ 'Guatemala': 'GT',
+ 'Guam': 'GU',
+ 'Guinea-Bissau': 'GW',
+ 'Guyana': 'GY',
+ 'Hong Kong': 'HK',
+ 'Heard Island and McDonald Islands': 'HM',
+ 'Heard Island And McDonald Islands': 'HM',
+ 'Honduras': 'HN',
+ 'Croatia': 'HR',
+ 'Haiti': 'HT',
+ 'Hungary': 'HU',
+ 'Indonesia': 'ID',
+ 'Ireland': 'IE',
+ 'Israel': 'IL',
+ 'Isle of Man': 'IM',
+ 'India': 'IN',
+ 'British Indian Ocean Territory': 'IO',
+ 'Iraq': 'IQ',
+ 'Iran': 'IR',
+ 'Iran (Islamic Republic of)': 'IR',
+ 'Iceland': 'IS',
+ 'Italy': 'IT',
+ 'Jersey': 'JE',
+ 'Jamaica': 'JM',
+ 'Jordan': 'JO',
+ 'Japan': 'JP',
+ 'Kenya': 'KE',
+ 'Kyrgyzstan': 'KG',
+ 'Cambodia': 'KH',
+ 'Kiribati': 'KI',
+ 'Comoros': 'KM',
+ 'Saint Kitts and Nevis': 'KN',
+ 'Saint Kitts And Nevis': 'KN',
+ 'North Korea': 'KP',
+ "Korea (Democratic People's Republic of)": 'KP',
+ 'South Korea': 'KR',
+ 'Korea, Republic of': 'KR',
+ 'Kuwait': 'KW',
+ 'Cayman Islands': 'KY',
+ 'Kazakhstan': 'KZ',
+ 'Laos': 'LA',
+ "Lao People's Democratic Republic": 'LA',
+ 'Lebanon': 'LB',
+ 'Saint Lucia': 'LC',
+ 'Liechtenstein': 'LI',
+ 'Sri Lanka': 'LK',
+ 'Liberia': 'LR',
+ 'Lesotho': 'LS',
+ 'Lithuania': 'LT',
+ 'Luxembourg': 'LU',
+ 'Latvia': 'LV',
+ 'Libya': 'LY',
+ 'Saint Martin (French part)': None,
+ 'Morocco': 'MA',
+ 'Monaco': 'MC',
+ 'Moldova': 'MD',
+ 'Moldova, Republic of': 'MD',
+ 'Montenegro': 'ME',
+ 'Madagascar': 'MG',
+ 'Marshall Islands': 'MH',
+ 'Macedonia': 'MK',
+ 'North Macedonia': 'MK',
+ 'Mali': 'ML',
+ 'Myanmar': 'MM',
+ 'Mongolia': 'MN',
+ 'Macau': 'MO',
+ 'Macao': 'MO',
+ 'Northern Mariana Islands': 'MP',
+ 'Martinique': 'MQ',
+ 'Mauritania': 'MR',
+ 'Montserrat': 'MS',
+ 'Malta': 'MT',
+ 'Mauritius': 'MU',
+ 'Maldives': 'MV',
+ 'Malawi': 'MW',
+ 'Mexico': 'MX',
+ 'Netherlands': 'NL',
+ 'Malaysia': 'MY',
+ 'Mozambique': 'MZ',
+ 'Namibia': None,
+ 'New Caledonia': 'NC',
+ 'Niger': 'NE',
+ 'Norfolk Island': 'NF',
+ 'Nigeria': 'NG',
+ 'Nicaragua': 'NI',
+ 'Norway': 'NO',
+ 'Nepal': 'NP',
+ 'Nauru': 'NR',
+ 'Niue': 'NU',
+ 'New Zealand': 'NZ',
+ 'Oman': 'OM',
+ 'Panama': 'PA',
+ 'Peru': 'PE',
+ 'French Polynesia': 'PF',
+ 'Papua New Guinea': 'PG',
+ 'Philippines': 'PH',
+ 'Pakistan': 'PK',
+ 'Poland': 'PL',
+ 'Pitcairn Islands': 'PN',
+ 'Pitcairn': 'PN',
+ 'Puerto Rico': 'PR',
+ 'Palestine': 'PS',
+ 'Palestine, State of': 'PS',
+ 'Portugal': 'PT',
+ 'Palau': 'PW',
+ 'Paraguay': 'PY',
+ 'Qatar': 'QA',
+ 'Réunion': 'RE',
+ 'Romania': 'RO',
+ 'Serbia': 'RS',
+ 'Russia': 'RU',
+ 'Russian Federation': 'RU',
+ 'Rwanda': 'RW',
+ 'Saudi Arabia': 'SA',
+ 'Solomon Islands': 'SB',
+ 'Seychelles': 'SC',
+ 'Sudan': 'SD',
+ 'Sweden': 'SE',
+ 'Singapore': 'SG',
+ 'Saint Helena': 'SH',
+ 'Saint Helena, Ascension and Tristan da Cunha': None,
+ 'Slovenia': 'SI',
+ 'Svalbard and Jan Mayen': 'SJ',
+ 'Svalbard And Jan Mayen': 'SJ',
+ 'Slovakia': 'SK',
+ 'Sierra Leone': 'SL',
+ 'San Marino': 'SM',
+ 'Senegal': 'SN',
+ 'Somalia': 'SO',
+ 'Saint Pierre and Miquelon': 'FR_PM',
+ 'Suriname': 'SR',
+ 'South Sudan': 'SS',
+ 'São Tomé and Príncipe': 'ST',
+ 'Sao Tome and Principe': 'ST',
+ 'São Tomé And Príncipe': 'ST',
+ 'El Salvador': 'SV',
+ 'Sint Maarten': 'SX',
+ 'Sint Maarten (Dutch part)': 'SX',
+ 'Syria': 'SY',
+ 'Syrian Arab Republic': 'SY',
+ 'Swaziland': 'SZ',
+ 'Eswatini': 'SZ',
+ 'Turks and Caicos Islands': 'TC',
+ 'Turks And Caicos Islands': 'TC',
+ 'Chad': 'TD',
+ 'French Southern Territories': 'TF',
+ 'Togo': 'TG',
+ 'Thailand': 'TH',
+ 'Tajikistan': 'TJ',
+ 'Tokelau': 'TK',
+ 'East Timor': 'TL',
+ 'Timor-Leste': 'TL',
+ 'Turkmenistan': 'TM',
+ 'Tunisia': 'TN',
+ 'Tonga': 'TO',
+ 'Turkey': 'TR',
+ 'Türkiye': 'TR',
+ 'Trinidad and Tobago': 'TT',
+ 'Tuvalu': 'TV',
+ 'Taiwan': 'TW',
+ 'Taiwan, Province of China': 'TW',
+ 'Tanzania': 'TZ',
+ 'Tanzania, United Republic of': 'TZ',
+ 'Ukraine': 'UA',
+ 'Uganda': 'UG',
+ 'United States Minor Outlying Islands': 'UM',
+ 'United States of America': 'US',
+ 'United States': 'US',
+ 'Kosovo': 'XK',
+ 'Yemen': 'YE',
+ 'Mayotte': 'YT',
+ 'South Africa': 'ZA',
+ 'Uruguay': 'UY',
+ 'Uzbekistan': 'UZ',
+ 'Vatican City': 'VA',
+ 'Holy See': 'VA',
+ 'Saint Vincent and the Grenadines': 'VC',
+ 'Saint Vincent And the Grenadines': 'VC',
+ 'Venezuela': 'VE',
+ 'Venezuela (Bolivarian Republic of)': 'VE',
+ 'British Virgin Islands': 'VG',
+ 'Virgin Islands (British)': 'VG',
+ 'United States Virgin Islands': 'VI',
+ 'Virgin Islands (U.S.)': 'VI',
+ 'Vietnam': 'VN',
+ 'Viet Nam': 'VN',
+ 'Viet-Nam': 'VN',
+ 'Vanuatu': 'VU',
+ 'Wallis and Futuna': 'WF',
+ 'Samoa': 'WS',
+ 'Zambia': 'ZM',
+ 'Zimbabwe': 'ZW'}
