+++ pyproject.toml
@@ -1,9 +1,10 @@
 [tool.poetry]
-name = "streamlit-docker-template"
+name = "mpx-dashboard"
 version = "0.1.0"
 description = ""
 authors = ["Flávio Codeço Coelho <fccoelho@gmail.com>"]
 license = "MIT"
+packages = [{include='mpx_dashboard'}]
 
 [tool.poetry.dependencies]
 python = "^3.9"
@@ -19,3 +20,4 @@ plotly = "^5.10.0"
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
